+++ android/src/main/java/expo/modules/libvlcplayer/constants/MediaPlayerConstants.kt
@@ -1,8 +1,8 @@
 package expo.modules.libvlcplayer.constants
 
 object MediaPlayerConstants {
-    const val DEFAULT_PLAYER_SCALE: Float = 0f
-    const val DEFAULT_PLAYER_RATE: Float = 1f
+    const val DEFAULT_PLAYER_SCALE: Double = 0.0
+    const val DEFAULT_PLAYER_RATE: Double = 1.0
     const val DEFAULT_PLAYER_TIME: Int = 0
     const val MIN_PLAYER_VOLUME: Int = 0
     const val MAX_PLAYER_VOLUME: Int = 100
