+++ ios/Records/MediaInfo.swift
@@ -8,7 +8,7 @@ struct MediaInfo: Record {
     var height: Int = 0
 
     @Field
-    var length: Double = 0.0
+    var length: Int = 0
 
     @Field
     var seekable: Bool = false
